requested changes & moving utils
