+++ libs/ethers/src/lib/key-manager/+state/key-manager.query.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 import { QueryEntity } from '@datorama/akita';
 import { KeyManagerStore, KeyState } from './key-manager.store';
 import { map, filter, first } from 'rxjs/operators';
-import { Key } from '@blockframes/utils';
+import { Key } from '../../types';
 
 @Injectable({ providedIn: 'root' })
 export class KeyManagerQuery extends QueryEntity<KeyState, Key> {
