+++ libs/ethers/src/lib/key-manager/+state/key-manager.store.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { StoreConfig, EntityStore, EntityState, ActiveState } from '@datorama/akita';
-import { Key } from '@blockframes/utils';
 import { BehaviorSubject } from 'rxjs';
+import { Key } from '../../types';
 
 export interface KeyState extends EntityState<Key>, ActiveState<string> {}
