+++ libs/ethers/src/lib/wallet/components/wallet-import-key/wallet-import-key.component.ts
@@ -3,7 +3,7 @@ import { KeyManagerService, KeyManagerQuery } from "../../../key-manager/+state"
 import { WalletQuery } from "../../+state";
 import { Observable } from "rxjs";
 import { Router } from "@angular/router";
-import { Key } from "@blockframes/utils";
+import { Key } from "../../../types";
 
 enum steps {
   import,
