+++ libs/ethers/src/lib/wallet/pages/wallet-reveal-mnemonic/wallet-reveal-mnemonic.component.ts
@@ -2,8 +2,8 @@ import { Component, ChangeDetectionStrategy, OnInit } from "@angular/core";
 import { Wallet as EthersWallet } from "@ethersproject/wallet";
 import { ActivatedRoute } from "@angular/router";
 import { KeyManagerQuery, KeyManagerService } from "../../../key-manager/+state";
-import { Key } from "@blockframes/utils";
 import { Observable } from "rxjs";
+import { Key } from "../../../types";
 
 enum steps {
   password,
