+++ libs/material/src/lib/delivery/pages/delivery-list/delivery-list.component.ts
@@ -1,11 +1,11 @@
 import { ChangeDetectionStrategy, Component, OnInit, ViewChild, HostBinding } from '@angular/core';
-import { MovieQuery } from 'libs/movie/src/lib/movie/+state/movie.query';
+import { MovieQuery } from '@blockframes/movie/movie/+state/movie.query';
 import { DeliveryQuery, Delivery } from '../../+state';
 import { Router } from '@angular/router';
 import { MatSort } from '@angular/material/sort';
 import { Observable} from 'rxjs';
 import { Organization, OrganizationQuery } from '@blockframes/organization';
-import { Movie } from '@blockframes/movie';
+import { Movie } from '@blockframes/movie/movie/+state/movie.model';
 
 @Component({
   selector: 'delivery-list',
